allow sto swap wlan client, set hostname, timezone and keyboard via config ui

+++ raspberry/startup-check.sh
@@ -12,6 +12,7 @@ log "started"
 
 LAST_PASSWD=""
 LAST_MCS=""
+LAST_WLAN_CLIENT=""
 
 if [ -f $LAST ] ; then
     source $LAST
@@ -24,6 +25,7 @@ if [ ! -f $CONFIG ]; then
 fi
 source <(tr -d '\015' < $CONFIG)
 hasChanges=0
+needsReboot=0
 if [ "$AVNAV_PASSWD" != "" ] ; then
     if [ "$AVNAV_PASSWD" = "$LAST_PASSWD" ] ; then
         log "AVNAV_PASSWD is set but unchanged"
@@ -41,8 +43,80 @@ else
     log "AVNAV_PASSWD not set"
 fi
 
+if [ "$AVNAV_WLAN_CLIENT" != "$LAST_WLAN_CLIENT" ] ; then
+    log "AVNAV_WLAN_CLIENT=$AVNAV_WLAN_CLIENT"
+    hasChanges=1
+    needsReboot=1
+    LAST_WLAN_CLIENT="$AVNAV_WLAN_CLIENT"
+else
+    log "AVNAV_WLAN_CLIENT unchanged"
+fi
+
+if [ "$AVNAV_HOSTNAME" != "$LAST_HOSTNAME" -a "$AVNAV_HOSTNAME" != "" ] ; then
+    log "AVNAV_HOSTNAME is set to $AVNAV_HOSTNAME"
+    current=`cat /etc/hostname | tr -d " \t\n\r"`
+    if [ "$current" = "$AVNAV_HOSTNAME" ]; then
+        log "hostname already correctly set in /etc/hostname"
+    else
+        echo "$AVNAV_HOSTNAME" > /etc/hostname
+        sed -i "s/127.0.1.1.*$current/127.0.1.1\t$AVNAV_HOSTNAME/g" /etc/hosts
+        hasChanges=1
+        needsReboot=1
+    fi
+    LAST_HOSTNAME="$AVNAV_HOSTNAME"
+else
+    log "AVNAV_HOSTNAME unchanged"
+fi
+
+if [ "$AVNAV_KBLAYOUT" != "$LAST_KBLAYOUT" -o "$AVNAV_KBMODEL" != "$LAST_KBMODEL" ] ; then
+    if [ "$AVNAV_KBLAYOUT" != "" -a "$AVNAV_KBMODEL" != "" ]; then
+        log "AVNAV_KBLAYOUT=$AVNAV_KBLAYOUT, AVNAV_KBMODEL=$AVNAV_KBMODEL"
+        kbfile=/etc/default/keyboard
+        currentModel=`sed -n 's/^ *XKBMODEL *= *//p' $kbfile | tr -d '" \n\r'`
+        currentLayout=`sed -n 's/^ *XKBLAYOUT *= *//p' $kbfile | tr -d '" \n\r'`
+        if [ "$currentLayout" = "$AVNAV_KBLAYOUT" -a "$currentModel" = "$AVNAV_KBMODEL" ] ; then
+            log "keyboard already correctly set up in system"
+        else
+            log "set up keyboard"
+            hasChanges=1
+            echo "XKBMODEL=\"$AVNAV_KBMODEL\"" > $kbfile
+            echo "XKBLAYOUT=\"$AVNAV_KBLAYOUT\"" >> $kbfile
+            echo "XKBVARIANT=\"\"" >> $kbfile
+            echo "XKBOPTIONS=\"\"" >> $kbfile
+            dpkg-reconfigure -f noninteractive keyboard-configuration
+        fi
+        
+        LAST_KBLAYOUT="$AVNAV_KBLAYOUT"
+        LAST_KBMODEL="$AVNAV_KBMODEL"
+    fi    
+else
+    log "AVNAV_KBLAYOUT and AVNAV_KBMODEL unchanged"
+fi
+
+if [ "$AVNAV_TIMEZONE" != "$LAST_TIMEZONE" -a "$AVNAV_TIMEZONE" != "" ] ; then
+    log "AVNAV_TIMEZONE is set to $AVNAV_TIMEZONE"
+    currentTZ=`cat /etc/timezone | tr -d " \t\n\r"`
+    if [ "$currentTZ" = "$AVNAV_TIMEZONE" ] ; then
+        log "timezone already correctly set in system"
+    else
+        log "change time zone"
+        rm -f /etc/localtime
+        hasChanges=1
+        needsReboot=1
+        echo "$AVNAV_TIMEZONE" > /etc/timezone
+        dpkg-reconfigure -f noninteractive tzdata
+    fi
+    LAST_TIMEZONE="$AVNAV_TIMEZONE"
+else
+    log "timezone unchanged"
+fi
+
 runMcs=0
 if [ "$AVNAV_MCS" = "yes" ] ; then
+    if grep MCS_DO_NOT_DELETE /etc/modules > /dev/null ; then
+        log "must correct /etc/modules"
+        LAST_MCS=""
+    fi
     if [ "$LAST_MCS" = "yes" ] ; then
         log "AVNAV_MCS is set but unchanged"
     else
@@ -59,7 +133,6 @@ else
     log "AVNAV_MCS not enabled"    
 fi
 
-needsReboot=0
 if [ "$runMcs" = 1 ];then
     log "running $MCS_INSTALL"
     $MCS_INSTALL -r -p -c
@@ -78,10 +151,19 @@ if [ "$hasChanges" = 1 ]; then
     log "writing back $LAST"
     echo "LAST_MCS=$LAST_MCS" > $LAST
     echo "LAST_PASSWD='$LAST_PASSWD'" >> $LAST
+    echo "LAST_WLAN_CLIENT='$LAST_WLAN_CLIENT'" >> $LAST
+    echo "LAST_HOSTNAME='$LAST_HOSTNAME'" >> $LAST
+    echo "LAST_KBLAYOUT='$LAST_KBLAYOUT'" >> $LAST
+    echo "LAST_KBMODEL='$LAST_KBMODEL'" >> $LAST
+    echo "LAST_TIMEZONE='$LAST_TIMEZONE'" >> $LAST
     chmod 600 $LAST
 fi
 if [ $needsReboot = 1 ] ; then
+    echo "****rebooting now****"
     log "****rebooting now****"
+    if [ "$1" = "noreboot" ] ; then
+        exit 2
+    fi
     reboot
 fi
 exit 0
